+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.3-unofficial-mushymato.1](UNRELEASED)
+## [0.6.3-unofficial-mushymato.1](https://github.com/Mushymato/StardewUI/releases/tag/0.6.3-unofficial-mushymato.1)
 
 ### Added
 - New `IMenuController` property `OpenSound`, set the sound which plays when menu becomes active for the first time.
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - `ReflectionPropertyDescriptor` should use `MethodInfoExtensions.SafeCreateDelegate` to correctly handle property getters on value types.
 - `button` screen-read should use delegate method of updating.
 - `checkbox` screen-read should include label-text.
+- `drop-down` overlay screen-read inherit them from labels.
 
 ## [0.6.3-unofficial-mushymato.0](https://github.com/Mushymato/StardewUI/releases/tag/0.6.3-unofficial-mushymato.0) - WIP